+++ src/main/java/io/github/seanboyy/enchantmentsreloaded/util/Config.java
@@ -14,14 +14,14 @@ public class Config {
         CLIENT = specPair.getLeft();
     }
 
-    public static float curse_removal_durability_factor;
+    public static double curse_removal_durability_factor;
     public static boolean curse_removal_durability_rounding_style;
-    public static float enchantment_transfer_use_break_chance;
+    public static double enchantment_transfer_use_break_chance;
 
     public static void bakeConfig() {
-        curse_removal_durability_factor = (float)(double)CLIENT.CURSE_DURABILITY_FACTOR.get();
+        curse_removal_durability_factor = CLIENT.CURSE_DURABILITY_FACTOR.get();
         curse_removal_durability_rounding_style = CLIENT.CURSE_ROUNDING_STYLE.get();
-        enchantment_transfer_use_break_chance = (float)(double)CLIENT.BREAK_CHANCE.get();
+        enchantment_transfer_use_break_chance = CLIENT.BREAK_CHANCE.get();
     }
 
     public static final class ClientConfig {
